+++ src/main/java/com/varadha2k/ExampleApplication.java
@@ -1,0 +1,20 @@
+package com.varadha2k;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.boot.web.client.RestTemplateBuilder;
+import org.springframework.context.annotation.Bean;
+import org.springframework.web.client.RestTemplate;
+
+@SpringBootApplication
+public class ExampleApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(ExampleApplication.class, args);
+	}
+
+	@Bean
+	public RestTemplate restTemplate() {
+		return new RestTemplateBuilder().build();
+	}
+}
