+++ src/main/java/com/varadha2k/ExampleController.java
@@ -1,0 +1,47 @@
+package com.varadha2k;
+
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.RestController;
+
+import com.varadha2k.person.Person;
+import com.varadha2k.person.PersonRepository;
+import com.varadha2k.weather.WeatherClient;
+import com.varadha2k.weather.WeatherResponse;
+
+import java.util.Optional;
+
+@RestController
+public class ExampleController {
+
+    private final PersonRepository personRepository;
+    private final WeatherClient weatherClient;
+
+    @Autowired
+    public ExampleController(final PersonRepository personRepository, final WeatherClient weatherClient) {
+        this.personRepository = personRepository;
+        this.weatherClient = weatherClient;
+    }
+
+    @GetMapping("/hello")
+    public String hello() {
+        return "Hello World!";
+    }
+
+    @GetMapping("/hello/{lastName}")
+    public String hello(@PathVariable final String lastName) {
+        Optional<Person> foundPerson = personRepository.findByLastName(lastName);
+
+        return foundPerson
+                .map(person -> String.format("Hello %s %s!", person.getFirstName(), person.getLastName()))
+                .orElse(String.format("Who is this '%s' you're talking about?", lastName));
+    }
+
+    @GetMapping("/weather")
+    public String weather() {
+        return weatherClient.fetchWeather()
+                .map(WeatherResponse::getSummary)
+                .orElse("Sorry, I couldn't fetch the weather for you :(");
+    }
+}
