+++ src/main/java/com/varadha2k/person/PersonRepository.java
@@ -1,0 +1,11 @@
+package com.varadha2k.person;
+
+import org.springframework.data.repository.CrudRepository;
+
+import java.util.Optional;
+
+public interface PersonRepository extends CrudRepository<Person, String> {
+
+    Optional<Person> findByLastName(String lastName);
+
+}
