+++ src/test/java/com/varadha2k/ExampleControllerAPITest.java
@@ -1,0 +1,63 @@
+package com.varadha2k;
+
+import org.junit.Test;
+import org.junit.runner.RunWith;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
+import org.springframework.boot.test.mock.mockito.MockBean;
+import org.springframework.test.context.junit4.SpringRunner;
+import org.springframework.test.web.servlet.MockMvc;
+
+import com.varadha2k.ExampleController;
+import com.varadha2k.person.Person;
+import com.varadha2k.person.PersonRepository;
+import com.varadha2k.weather.WeatherClient;
+import com.varadha2k.weather.WeatherResponse;
+
+import java.util.Optional;
+
+import static org.mockito.BDDMockito.given;
+import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
+import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.content;
+import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
+
+@RunWith(SpringRunner.class)
+@WebMvcTest(controllers = ExampleController.class)
+public class ExampleControllerAPITest {
+
+    @Autowired
+    private MockMvc mockMvc;
+
+    @MockBean
+    private PersonRepository personRepository;
+
+    @MockBean
+    private WeatherClient weatherClient;
+
+    @Test
+    public void shouldReturnHelloWorld() throws Exception {
+        mockMvc.perform(get("/hello"))
+                .andExpect(content().string("Hello World!"))
+                .andExpect(status().is2xxSuccessful());
+    }
+
+    @Test
+    public void shouldReturnFullName() throws Exception {
+        Person peter = new Person("Peter", "Pan");
+        given(personRepository.findByLastName("Pan")).willReturn(Optional.of(peter));
+
+        mockMvc.perform(get("/hello/Pan"))
+                .andExpect(content().string("Hello Peter Pan!"))
+                .andExpect(status().is2xxSuccessful());
+    }
+
+    @Test
+    public void shouldReturnCurrentWeather() throws Exception {
+        WeatherResponse weatherResponse = new WeatherResponse("Hamburg, 8°C raining");
+        given(weatherClient.fetchWeather()).willReturn(Optional.of(weatherResponse));
+
+        mockMvc.perform(get("/weather"))
+                .andExpect(status().is2xxSuccessful())
+                .andExpect(content().string("Hamburg, 8°C raining"));
+    }
+}
