+++ src/test/java/com/varadha2k/HelloE2ESeleniumTest.java
@@ -1,0 +1,52 @@
+package com.varadha2k;
+
+import io.github.bonigarcia.wdm.ChromeDriverManager;
+import org.junit.After;
+import org.junit.Before;
+import org.junit.BeforeClass;
+import org.junit.Test;
+import org.junit.runner.RunWith;
+import org.openqa.selenium.By;
+import org.openqa.selenium.WebDriver;
+import org.openqa.selenium.WebElement;
+import org.openqa.selenium.chrome.ChromeDriver;
+import org.springframework.boot.context.embedded.LocalServerPort;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.test.context.junit4.SpringRunner;
+
+import static org.hamcrest.CoreMatchers.containsString;
+import static org.hamcrest.MatcherAssert.assertThat;
+
+@RunWith(SpringRunner.class)
+@SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
+public class HelloE2ESeleniumTest {
+
+    private WebDriver driver;
+
+    @LocalServerPort
+    private int port;
+
+    @BeforeClass
+    public static void setUpClass() throws Exception {
+        ChromeDriverManager.getInstance().setup();
+    }
+
+    @Before
+    public void setUp() throws Exception {
+        driver = new ChromeDriver();
+    }
+
+    @After
+    public void tearDown(){
+        driver.close();
+    }
+
+    @Test
+    public void helloPageHasTextHelloWorld(){
+        driver.navigate().to(String.format("http://localhost:%s/hello", port));
+
+        WebElement body = driver.findElement(By.tagName("body"));
+
+        assertThat(body.getText(), containsString("Hello World!"));
+    }
+}
