+++ src/test/java/com/varadha2k/helper/FileLoader.java
@@ -1,0 +1,14 @@
+package com.varadha2k.helper;
+
+import org.springframework.util.ResourceUtils;
+
+import java.io.File;
+import java.io.IOException;
+import java.nio.file.Files;
+
+public class FileLoader {
+    public static String read(String filePath) throws IOException {
+        File file = ResourceUtils.getFile(filePath);
+        return new String(Files.readAllBytes(file.toPath()));
+    }
+}
