+++ src/test/java/com/varadha2k/weather/WeatherAcceptanceTest.java
@@ -1,0 +1,45 @@
+package com.varadha2k.weather;
+
+import com.github.tomakehurst.wiremock.junit.WireMockRule;
+import com.varadha2k.helper.FileLoader;
+
+import org.junit.Rule;
+import org.junit.Test;
+import org.junit.runner.RunWith;
+import org.springframework.boot.context.embedded.LocalServerPort;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.http.MediaType;
+import org.springframework.test.context.junit4.SpringRunner;
+
+import static com.github.tomakehurst.wiremock.client.WireMock.*;
+import static io.restassured.RestAssured.when;
+import static org.hamcrest.Matchers.containsString;
+import static org.hamcrest.core.Is.is;
+import static org.springframework.http.HttpHeaders.CONTENT_TYPE;
+
+@RunWith(SpringRunner.class)
+@SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
+public class WeatherAcceptanceTest {
+
+    @LocalServerPort
+    private int port;
+
+    @Rule
+    public WireMockRule wireMockRule = new WireMockRule(8089);
+
+
+    @Test
+    public void shouldReturnYesterdaysWeather() throws Exception {
+        wireMockRule.stubFor(get(urlPathEqualTo("/some-test-api-key/53.5511,9.9937"))
+                .willReturn(aResponse()
+                        .withBody(FileLoader.read("classpath:weatherApiResponse.json"))
+                        .withHeader(CONTENT_TYPE, MediaType.APPLICATION_JSON_VALUE)
+                        .withStatus(200)));
+
+        when()
+                .get(String.format("http://localhost:%s/weather", port))
+                .then()
+                .statusCode(is(200))
+                .body(containsString("Rain"));
+    }
+}
