+++ src/test/java/com/varadha2k/weather/WeatherResponseTest.java
@@ -1,0 +1,23 @@
+package com.varadha2k.weather;
+
+import com.fasterxml.jackson.databind.ObjectMapper;
+import com.varadha2k.helper.FileLoader;
+import com.varadha2k.weather.WeatherResponse;
+
+import org.junit.Test;
+
+import static org.hamcrest.core.Is.is;
+import static org.junit.Assert.assertThat;
+
+public class WeatherResponseTest {
+
+    @Test
+    public void shouldDeserializeJson() throws Exception {
+        String jsonResponse = FileLoader.read("classpath:weatherApiResponse.json");
+        WeatherResponse expectedResponse = new WeatherResponse("Rain");
+
+        WeatherResponse parsedResponse = new ObjectMapper().readValue(jsonResponse, WeatherResponse.class);
+
+        assertThat(parsedResponse, is(expectedResponse));
+    }
+}
