--- src/test/java/example/helper/FileLoader.java
@@ -1,14 +1,0 @@
-package example.helper;
-
-import org.springframework.util.ResourceUtils;
-
-import java.io.File;
-import java.io.IOException;
-import java.nio.file.Files;
-
-public class FileLoader {
-    public static String read(String filePath) throws IOException {
-        File file = ResourceUtils.getFile(filePath);
-        return new String(Files.readAllBytes(file.toPath()));
-    }
-}
