--- src/test/java/example/person/PersonRepositoryIntegrationTest.java
@@ -1,36 +1,0 @@
-package example.person;
-
-import org.junit.After;
-import org.junit.Test;
-import org.junit.runner.RunWith;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest;
-import org.springframework.test.context.junit4.SpringRunner;
-
-import java.util.Optional;
-
-import static org.hamcrest.core.Is.is;
-import static org.junit.Assert.assertThat;
-
-@RunWith(SpringRunner.class)
-@DataJpaTest
-public class PersonRepositoryIntegrationTest {
-
-    @Autowired
-    private PersonRepository subject;
-
-    @After
-    public void tearDown() throws Exception {
-        subject.deleteAll();
-    }
-
-    @Test
-    public void shouldSaveAndFetchPerson() throws Exception {
-        Person peter = new Person("Peter", "Pan");
-        subject.save(peter);
-
-        Optional<Person> maybePeter = subject.findByLastName("Pan");
-
-        assertThat(maybePeter, is(Optional.of(peter)));
-    }
-}
