--- src/test/java/example/weather/WeatherClientIntegrationTest.java
@@ -1,43 +1,0 @@
-package example.weather;
-
-import com.github.tomakehurst.wiremock.junit.WireMockRule;
-import example.helper.FileLoader;
-import org.junit.Rule;
-import org.junit.Test;
-import org.junit.runner.RunWith;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.http.MediaType;
-import org.springframework.test.context.junit4.SpringRunner;
-
-import java.util.Optional;
-
-import static com.github.tomakehurst.wiremock.client.WireMock.*;
-import static org.hamcrest.MatcherAssert.assertThat;
-import static org.hamcrest.core.Is.is;
-import static org.springframework.http.HttpHeaders.CONTENT_TYPE;
-
-@RunWith(SpringRunner.class)
-@SpringBootTest
-public class WeatherClientIntegrationTest {
-
-    @Autowired
-    private WeatherClient subject;
-
-    @Rule
-    public WireMockRule wireMockRule = new WireMockRule(8089);
-
-    @Test
-    public void shouldCallWeatherService() throws Exception {
-        wireMockRule.stubFor(get(urlPathEqualTo("/some-test-api-key/53.5511,9.9937"))
-                .willReturn(aResponse()
-                        .withBody(FileLoader.read("classpath:weatherApiResponse.json"))
-                        .withHeader(CONTENT_TYPE, MediaType.APPLICATION_JSON_VALUE)
-                        .withStatus(200)));
-
-        Optional<WeatherResponse> weatherResponse = subject.fetchWeather();
-
-        Optional<WeatherResponse> expectedResponse = Optional.of(new WeatherResponse("Rain"));
-        assertThat(weatherResponse, is(expectedResponse));
-    }
-}
